refactor(app/fireworks): break the sparks out into thier own module
