+++ src/components/Fireworks/Fireworks.tsx
@@ -1,71 +1,14 @@
 import React, { Dispatch, FC, useEffect, useRef, useState } from 'react';
 
 import { AppAction } from '../../app';
-import {
-  add,
-  caf,
-  cos,
-  dpr,
-  Particle,
-  raf,
-  random,
-  sin,
-  sub,
-  Vec2,
-  π,
-  ππ,
-  pool,
-} from '../../lib';
+import { caf, dpr, raf, random } from '../../lib';
 
 import style from './Fireworks.module.css';
+import { activate, numToActivate, render, sparks, update } from './particles';
 
 interface Props {
   dispatch: Dispatch<AppAction>;
 }
-
-interface Spark extends Particle {
-  hue: number;
-}
-
-const maxInitialSpeed = 15;
-const gravity: Vec2 = { x: 0, y: 0.2 };
-
-const sparks = pool<Spark>(5_000, {
-  hue: 0,
-});
-
-const activateSpark: (spark: Spark, x: number, y: number) => void = (
-  spark,
-  x,
-  y,
-) => {
-  const theta = random() * ππ;
-  const radius = random() * maxInitialSpeed;
-
-  spark.currPos = { x, y };
-  spark.prevPos = { x: x + cos(theta) * radius, y: y + sin(theta) * radius };
-  spark.active = true;
-  spark.hue = random() * ππ * (180 / π);
-};
-
-const updateSpark: (spark: Spark) => void = (spark) => {
-  const currVel = add(sub(spark.currPos, spark.prevPos), gravity);
-  const nextPos = add(spark.currPos, currVel);
-
-  spark.prevPos = spark.currPos;
-  spark.currPos = nextPos;
-};
-
-const renderSpark: (context: CanvasRenderingContext2D, spark: Spark) => void = (
-  ctx,
-  { currPos, prevPos, hue },
-) => {
-  ctx.fillStyle = 'white';
-  ctx.fillRect(prevPos.x - 4, prevPos.y - 4, 8, 8);
-
-  ctx.fillStyle = `hsl(${hue}, 80%, 50%)`;
-  ctx.fillRect(currPos.x - 4, currPos.y - 4, 8, 8);
-};
 
 export const Fireworks: FC<Props> = ({ dispatch }) => {
   const fireworksElRef = useRef<HTMLCanvasElement>(null);
@@ -110,16 +53,16 @@ export const Fireworks: FC<Props> = ({ dispatch }) => {
       context.fillStyle = 'black';
       context.fillRect(0, 0, width, height);
 
-      let i = 0;
+      let numActivated = 0;
       const shouldEmit = time % 1_000 < 18 && time < 5_000;
       const emitX = hw / 2 + random() * hw;
       const emitY = hh / 4 + random() * hh;
 
       sparks.forEach((spark) => {
         if (!spark.active && shouldEmit) {
-          if (i < 500) {
-            activateSpark(spark, emitX, emitY);
-            ++i;
+          if (numActivated < numToActivate) {
+            activate(spark, emitX, emitY);
+            ++numActivated;
           } else {
             return;
           }
@@ -135,8 +78,8 @@ export const Fireworks: FC<Props> = ({ dispatch }) => {
           return;
         }
 
-        updateSpark(spark);
-        renderSpark(context, spark);
+        update(spark);
+        render(context, spark);
       });
 
       if (
